swing-storage-tooling/requirements.txt
supabase
boto3
python-dotenv
